refactored package structure - classes packaged by related entity instead of model, repository, and service
