+++ src/main/java/com/storyblocks/storyblocksservice/users/UsersRepository.java
@@ -1,6 +1,6 @@
-package com.storyblocks.storyblocksservice.repository;
+package com.storyblocks.storyblocksservice.users;
 
-import com.storyblocks.storyblocksservice.model.User;
+import com.storyblocks.storyblocksservice.users.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
