+++ src/test/java/com/storyblocks/storyblocksservice/users/RegistrationServiceTests.java
@@ -1,11 +1,11 @@
-package com.storyblocks.storyblocksservice.service;
+package com.storyblocks.storyblocksservice.users;
 
-import com.storyblocks.storyblocksservice.controller.RegistrationRequest;
-import com.storyblocks.storyblocksservice.model.ConfirmationToken;
-import com.storyblocks.storyblocksservice.model.User;
-import com.storyblocks.storyblocksservice.service.impl.ConfirmationTokenService;
-import com.storyblocks.storyblocksservice.service.impl.JpaUserDetailsManager;
-import com.storyblocks.storyblocksservice.service.impl.RegistrationService;
+import com.storyblocks.storyblocksservice.users.RegistrationRequest;
+import com.storyblocks.storyblocksservice.users.ConfirmationToken;
+import com.storyblocks.storyblocksservice.users.User;
+import com.storyblocks.storyblocksservice.users.ConfirmationTokenService;
+import com.storyblocks.storyblocksservice.users.JpaUserDetailsManager;
+import com.storyblocks.storyblocksservice.users.RegistrationService;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
