+++ src/test/java/com/storyblocks/storyblocksservice/users/UserDetailsManagerTests.java
@@ -1,10 +1,10 @@
-package com.storyblocks.storyblocksservice.service;
+package com.storyblocks.storyblocksservice.users;
 
-import com.storyblocks.storyblocksservice.model.ConfirmationToken;
-import com.storyblocks.storyblocksservice.model.User;
-import com.storyblocks.storyblocksservice.repository.UsersRepository;
-import com.storyblocks.storyblocksservice.service.impl.ConfirmationTokenService;
-import com.storyblocks.storyblocksservice.service.impl.JpaUserDetailsManager;
+import com.storyblocks.storyblocksservice.users.ConfirmationToken;
+import com.storyblocks.storyblocksservice.users.User;
+import com.storyblocks.storyblocksservice.users.UsersRepository;
+import com.storyblocks.storyblocksservice.users.ConfirmationTokenService;
+import com.storyblocks.storyblocksservice.users.JpaUserDetailsManager;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
